Moved errors and constants to their own files to avoid circular dependencies. Move scripting to parlay.scripts. Added import to parlay.utils to maintain backwards compatibility, but this will be deprecated soon
