+++ parlay/constants.py
@@ -1,0 +1,2 @@
+DEFAULT_TIMEOUT = 600  # unless modified, waiting for any message will raise TimeoutError after this time
+
