+++ parlay/errors.py
@@ -1,0 +1,3 @@
+
+class TimeoutError(Exception):
+    pass
